`VerticalTimetableView`のファイル分割を実施
